+++ src/main/java/com/fishcount/common/model/entity/TituloParcela.java
@@ -23,6 +23,7 @@ import javax.persistence.TemporalType;
 import com.fishcount.common.model.enums.EnumStatusTitulo;
 import com.fishcount.common.model.enums.EnumTipoTitulo;
 import com.fishcount.common.model.pattern.AbstractEntity;
+import java.math.BigDecimal;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -44,16 +45,16 @@ public class TituloParcela extends AbstractEntity<Integer> {
     private Integer id;
 
     @Column(name = "valor")
-    private Double valor;
+    private BigDecimal valor;
 
     @Column(name = "saldo")
-    private Double saldo;
+    private BigDecimal saldo;
 
     @Column(name = "acrescimo")
-    private Double acrescimo;
+    private BigDecimal acrescimo;
 
     @Column(name = "desconto")
-    private Double desconto;
+    private BigDecimal desconto;
 
     @Column(name = "data_vencimento")
     @Temporal(TemporalType.TIMESTAMP)
